feat(www): let the preset preview switch scenes
Adds an App / Style guide / Cards switcher to the picker's preview pane,
rendering the /create scenes inline (lazily, no iframe) and deferring the
previewed system so the rail keeps up with the keyboard.

Also: no gap between rail and preview, close button outside the panel, no
border under the search field, and the rail now builds on Autocomplete +
ListBox directly instead of the Command wrapper it had to fight.
